Renames component name for new page
