+++ src/app/app.module.ts
@@ -29,7 +29,7 @@ import { LoggingService } from './core/services/logging.service';
 import { PhaseService } from './core/services/phase.service';
 import { SessionFixConfirmationComponent } from './core/services/session-fix-confirmation/session-fix-confirmation.component';
 import { UserService } from './core/services/user.service';
-import { DetailedViewerModule } from './detailed-viewer/detailed-viewer.module';
+import { DetailedViewerModule } from './detailed-viewer/user-detail-viewer.module';
 import { IssuesViewerModule } from './issues-viewer/issues-viewer.module';
 import { LabelDefinitionPopupComponent } from './shared/label-definition-popup/label-definition-popup.component';
 import { HeaderComponent } from './shared/layout';
