+++ src/app/detailed-viewer/user-detail-viewer-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from '../core/guards/auth.guard';
-import { DetailedViewerComponent } from './detailed-viewer.component';
+import { UserDetailViewerComponent } from './user-detail-viewer.component';
 
-const routes: Routes = [{ path: 'user/:name', component: DetailedViewerComponent, canActivate: [AuthGuard] }];
+const routes: Routes = [{ path: 'user/:name', component: UserDetailViewerComponent, canActivate: [AuthGuard] }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
