+++ src/app/detailed-viewer/user-detail-viewer.component.css
